+++ Sources/DoggieGraphics/ImageContext/SetClip.swift
@@ -72,7 +72,7 @@ extension ImageContext {
                     _stencil += antialias * _min_x + _min_y * _stencil_width2
                     
                     let _antialias2 = antialias * antialias
-                    let div = Double(_antialias2)
+                    let div = Float(_antialias2)
                     
                     for _ in _min_y..<_max_y {
                         
@@ -97,7 +97,7 @@ extension ImageContext {
                             }
                             
                             if _p != 0 {
-                                _clip.pointee = _p == _antialias2 ? 1 : Double(_p) / div
+                                _clip.pointee = _p == _antialias2 ? 1 : Float(_p) / div
                             }
                             
                             _clip += 1
